tech(api): move models from scoring to evaluation bounded context
