+++ api/src/certification/evaluation/domain/models/CertificationAssessmentHistory.js
@@ -1,5 +1,5 @@
 /**
- * @typedef {import('../../../evaluation/domain/models/FlashAssessmentAlgorithm.js').FlashAssessmentAlgorithm} FlashAssessmentAlgorithm
+ * @typedef {import('./FlashAssessmentAlgorithm.js').FlashAssessmentAlgorithm} FlashAssessmentAlgorithm
  * @typedef {import('../read-models/ChallengeCalibration.js').ChallengeCalibration} ChallengeCalibration
  * @typedef {import('../../../../evaluation/domain/models/Answer.js').Answer} Answer
  */
