+++ api/src/certification/evaluation/domain/models/CertificationAssessmentScore.js
