+++ api/src/certification/evaluation/domain/models/CertificationAssessmentScoreV3.js
@@ -11,8 +11,8 @@ import { config } from '../../../../shared/config.js';
 import { COMPETENCES_COUNT, PIX_COUNT_BY_LEVEL } from '../../../../shared/domain/constants.js';
 import { status as CertificationStatus } from '../../../../shared/domain/models/AssessmentResult.js';
 import { meshConfiguration } from '../../../results/domain/models/v3/MeshConfiguration.js';
-import { Intervals } from '../../../scoring/domain/models/Intervals.js';
 import { ABORT_REASONS } from '../../../shared/domain/constants/abort-reasons.js';
+import { Intervals } from './Intervals.js';
 
 export class CertificationAssessmentScoreV3 {
   /**
