+++ api/src/certification/evaluation/domain/models/Intervals.js
