+++ api/src/certification/evaluation/domain/models/ScoringAndCapacitySimulatorReport.js
