+++ api/src/certification/evaluation/domain/models/ScoringSimulator.js
