+++ api/src/certification/evaluation/domain/services/scoring/create-v3-assessment-result.js
@@ -1,5 +1,5 @@
 import { config } from '../../../../../shared/config.js';
-import { AssessmentResultFactory } from '../../../../scoring/domain/models/factories/AssessmentResultFactory.js';
+import { AssessmentResultFactory } from '../../models/factories/AssessmentResultFactory.js';
 
 export function createV3AssessmentResult({
   toBeCancelled,
