+++ api/tests/certification/evaluation/integration/infrastructure/repositories/challenge-calibration-repository_test.js
@@ -1,5 +1,5 @@
+import { ChallengeCalibration } from '../../../../../../src/certification/evaluation/domain/read-models/ChallengeCalibration.js';
 import { getByCertificationCourseId } from '../../../../../../src/certification/evaluation/infrastructure/repositories/challenge-calibration-repository.js';
-import { ChallengeCalibration } from '../../../../../../src/certification/scoring/domain/read-models/ChallengeCalibration.js';
 import { databaseBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Integration | Infrastructure | Repository | ChallengeCalibrationRepository', function () {
