+++ api/tests/certification/evaluation/unit/domain/models/CertificationAssessmentHistory_test.js
@@ -1,7 +1,7 @@
-import { CertificationAssessmentHistory } from '../../../../../../src/certification/scoring/domain/models/CertificationAssessmentHistory.js';
+import { CertificationAssessmentHistory } from '../../../../../../src/certification/evaluation/domain/models/CertificationAssessmentHistory.js';
 import { domainBuilder, expect, sinon } from '../../../../../test-helper.js';
 
-describe('Unit | Domain | Models | CertificationAssessmentHistory', function () {
+describe('Certification | Evaluation | Unit | Domain | Models | CertificationAssessmentHistory', function () {
   describe('#fromChallengesAndAnswers', function () {
     it('should return a CertificationAssessmentHistory with the capacity history', function () {
       // given
