+++ api/tests/certification/evaluation/unit/domain/models/CertificationAssessmentScore_test.js
@@ -1,7 +1,7 @@
-import { CertificationAssessmentScore } from '../../../../../../src/certification/scoring/domain/models/CertificationAssessmentScore.js';
+import { CertificationAssessmentScore } from '../../../../../../src/certification/evaluation/domain/models/CertificationAssessmentScore.js';
 import { domainBuilder, expect } from '../../../../../test-helper.js';
 
-describe('Unit | Domain | Models | CertificationAssessmentScore', function () {
+describe('Certification | Evaluation | Unit | Domain | Models | CertificationAssessmentScore', function () {
   describe('#get nbPix', function () {
     context('when CertificationAssessmentScore has no competence marks', function () {
       it('should return nbPix 0', function () {
