+++ api/tests/certification/evaluation/unit/domain/models/ScoringSimulator_test.js
@@ -1,7 +1,7 @@
-import { ScoringSimulator } from '../../../../../../src/certification/scoring/domain/models/ScoringSimulator.js';
+import { ScoringSimulator } from '../../../../../../src/certification/evaluation/domain/models/ScoringSimulator.js';
 import { domainBuilder, expect } from '../../../../../test-helper.js';
 
-describe('Unit | Domain | Models | ScoringSimulator', function () {
+describe('Certification | Evaluation | Unit | Domain | Models | ScoringSimulator', function () {
   describe('#compute', function () {
     [
       {
