+++ api/tests/certification/evaluation/unit/domain/models/factories/AssessmentResultFactory_test.js
@@ -1,9 +1,9 @@
-import { AssessmentResultFactory } from '../../../../../../src/certification/scoring/domain/models/factories/AssessmentResultFactory.js';
-import { AutoJuryCommentKeys } from '../../../../../../src/certification/shared/domain/models/JuryComment.js';
-import { AssessmentResult } from '../../../../../../src/shared/domain/models/AssessmentResult.js';
-import { domainBuilder, expect } from '../../../../../test-helper.js';
+import { AssessmentResultFactory } from '../../../../../../../src/certification/evaluation/domain/models/factories/AssessmentResultFactory.js';
+import { AutoJuryCommentKeys } from '../../../../../../../src/certification/shared/domain/models/JuryComment.js';
+import { AssessmentResult } from '../../../../../../../src/shared/domain/models/AssessmentResult.js';
+import { domainBuilder, expect } from '../../../../../../test-helper.js';
 
-describe('Certification | Scoring | Unit | Domain | Factories | AssessmentResultFactory', function () {
+describe('Certification | Evaluation | Unit | Domain | Factories | AssessmentResultFactory', function () {
   describe('#buildAlgoErrorResult', function () {
     it('should return an algo error AssessmentResult', function () {
       // given
