+++ api/tests/tooling/domain-builder/factory/build-certification-assessment-score.js
@@ -1,4 +1,4 @@
-import { CertificationAssessmentScore } from '../../../../src/certification/scoring/domain/models/CertificationAssessmentScore.js';
+import { CertificationAssessmentScore } from '../../../../src/certification/evaluation/domain/models/CertificationAssessmentScore.js';
 
 const buildCertificationAssessmentScore = function ({
   competenceMarks = [],
