+++ api/tests/tooling/domain-builder/factory/certification/scoring/build-certification-challenge-capacity.js
@@ -1,4 +1,4 @@
-import { CertificationChallengeCapacity } from '../../../../../../src/certification/scoring/domain/models/CertificationChallengeCapacity.js';
+import { CertificationChallengeCapacity } from '../../../../../../src/certification/evaluation/domain/models/CertificationChallengeCapacity.js';
 
 export const buildCertificationChallengeCapacity = ({ answerId, certificationChallengeId, capacity, createdAt }) => {
   return new CertificationChallengeCapacity({
